requirements.txt.txt
Flask
openai
python-docx
pytesseract
Pillow
pandas
opencv-python
PyMuPDF
ultralytics
flask-cors